Update Markup and Style Sheets
Added default markup and styling for Contact and About pages, as well as a huge revision on the whole markup and scss modules. Site is currently navigable through it's 3 main pages, with project page template likely coming up on the next update. Still no JS/AJAX, but that will be detailed in a roadmap text file that should come in between this and the next update.